QGIS refactor - kamangir/bolt#746
