+++ blue_geo/QGIS/console/apps/template.py
@@ -1,5 +1,5 @@
 if not QGIS_is_live:
-    from application import BLUE_GEO_QGIS_APPLICATION
+    from application import BLUE_GEO_QGIS_APPLICATION, Q_app_list
     from QGIS import QGIS
 
 
@@ -18,4 +18,4 @@ class BLUE_GEO_QGIS_APPLICATION_TEMPLATE(BLUE_GEO_QGIS_APPLICATION):
 
 
 template = BLUE_GEO_QGIS_APPLICATION_TEMPLATE()
-QGIS.add_application(template)
+Q_app_list.add(template)
