+++ blue_geo/QGIS/console/project.py
@@ -1,7 +1,9 @@
 import os
+from tqdm import tqdm
 
 if not QGIS_is_live:
     from .logger import Q_verbose, Q_log
+    from .graphics import Q_refresh
     from .file_load import Q_load_yaml
     from .mock import QgsProject
 
@@ -10,19 +12,49 @@ class ABCLI_QGIS_Project(object):
     def help(self):
         pass
 
+    def exists(
+        self,
+        layer_name: str,
+        verbose: bool = True,
+    ) -> bool:
+        list_of_layers = self.get_layer(layer_name=layer_name)
+
+        if len(list_of_layers) > 0:
+            if verbose:
+                Q_log(layer_name, icon="✅")
+            return True
+
+        return False
+
     @property
     def filename(self):
         return QgsProject.instance().fileName()
 
+    def get_layer(self, layer_name):
+        return QgsProject.instance().mapLayersByName(layer_name)
+
     @property
-    def list_of_layers(self):
+    def list_of_layers(
+        self,
+        aux: bool = False,
+    ):
         output = [layer.name() for layer in QgsProject.instance().mapLayers().values()]
+
+        if not aux:
+            output = [
+                layer_name
+                for layer_name in output
+                if not layer_name.startswith("Google")
+                and not layer_name.startswith("template")
+            ]
+
         if Q_verbose:
             Q_log(
                 f"{len(output)} layer(s)",
                 ", ".join(output),
                 icon="🔎",
             )
+
         return output
 
     @property
@@ -41,6 +73,25 @@ class ABCLI_QGIS_Project(object):
     @property
     def path(self):
         return QgsProject.instance().homePath()
+
+    def reload(self):
+        # https://gis.stackexchange.com/a/449101/210095
+        for layer_ in tqdm(QgsProject.instance().mapLayers().values()):
+            layer_.dataProvider().reloadData()
+
+    def remove_layer(
+        self,
+        layer_name: str,
+        refresh: bool = True,
+    ):
+        Q_log(layer_name, icon="🗑️")
+
+        for layer_ in QgsProject.instance().mapLayersByName(layer_name):
+            QgsProject.instance().removeMapLayer(layer_.id())
+            Q_log(layer_name, icon="➖")
+
+        if refresh:
+            Q_refresh()
 
     def trim_empty_groups(self):
         root = QgsProject.instance().layerTreeRoot()
@@ -62,4 +113,4 @@ class ABCLI_QGIS_Project(object):
                 )
 
 
-project = ABCLI_QGIS_Project()
+Q_project = ABCLI_QGIS_Project()
