+++ blue_geo/QGIS/console/tests/layer.py
@@ -1,6 +1,6 @@
 if not QGIS_is_live:
     from logger import Q_hr, Q_log, Q_log_error, Q_verbose
-    from ..layer import layer
+    from ..layer import Q_layer
 
 
 def test_layer_properties(
